[QPE-65] Moved over build pipeline files
