+++ .gitattributes
@@ -1,0 +1,11 @@
+*       text=auto
+.*      text eol=lf
+*.css   text eol=lf
+*.scss  text eol=lf
+*.html  text eol=lf
+*.js    text eol=lf
+*.json  text eol=lf
+*.md    text eol=lf
+*.sh    text eol=lf
+*.txt   text eol=lf
+*.svg   text eol=lf
