--- dpi-simple-table.qext
@@ -1,7 +1,0 @@
-{
-	"name": "Simple Table (w/ Links)",
-	"description": "",
-	"type": "visualization",
-	"version": 1,
-	"author": "Daniel Pilla"
-}
