+++ src/qlik-image-table.css
