+++ src/wbfolder.wbl
@@ -1,0 +1,3 @@
+qlik-image-table.qext;
+qlik-image-table.js;
+qlik-image-table.css
