--- wbfolder.wbl
@@ -1,3 +1,0 @@
-dpi-simple-table.qext;
-dpi-simple-table.js;
-dpi-simple-table.css
